refactor: Controller와 Converter 분리
