+++ docs/README.md
@@ -24,7 +24,7 @@
 
 ## 1 단계 피드백, 수정사항
 
-- [ ] Controller와 Converter는 앞에 문자도 겹치고 각 도메인별로 모두 있어서 헷갈릴 수 있을 것 같은데요, 패키지를 분리해 보면 어떨까요?
+- [x] Controller와 Converter는 앞에 문자도 겹치고 각 도메인별로 모두 있어서 헷갈릴 수 있을 것 같은데요, 패키지를 분리해 보면 어떨까요?
 - [x] (컨트롤러) 보통 Info라는 단어는 명확한 의미를 드러내지 못하는 경우가 많아 지양하는 것이 좋은데요, 여기서도 마찬가지인 것 같아요.
   - 코드에서 Controller가 요청을 받거나 응답을 주는 객체를 그대로 서비스에서 쓰지 않고 서비스와 컨트롤러 사이의 dto가 있습니다.
   - toInfo와 dto.info의 의미는 컨트롤러와 서비스 사이에 정보를 주고받는 dto를 의미합니다.
