+++ src/main/java/wooteco/subway/dto/converter/LineConverter.java
@@ -1,4 +1,4 @@
-package wooteco.subway.controller;
+package wooteco.subway.dto.converter;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -10,16 +10,16 @@ import wooteco.subway.dto.service.LineServiceResponse;
 import wooteco.subway.dto.service.LineUpdateRequest;
 
 public class LineConverter {
-    static LineUpdateRequest toServiceRequest(Long id, LineRequest lineRequest) {
+    public static LineUpdateRequest toServiceRequest(Long id, LineRequest lineRequest) {
         return new LineUpdateRequest(id, lineRequest.getName(), lineRequest.getColor(), lineRequest.getExtraFare());
     }
 
-    static LineServiceRequest toServiceRequest(LineRequest lineRequest) {
+    public static LineServiceRequest toServiceRequest(LineRequest lineRequest) {
         return new LineServiceRequest(lineRequest.getName(), lineRequest.getColor(), lineRequest.getUpStationId(),
             lineRequest.getDownStationId(), lineRequest.getDistance(), lineRequest.getExtraFare());
     }
 
-    static LineResponse toResponse(LineServiceResponse lineServiceResponse) {
+    public static LineResponse toResponse(LineServiceResponse lineServiceResponse) {
         List<StationResponse> stationResponses = lineServiceResponse.getStationInfos().stream()
             .map(StationConverter::toResponse)
             .collect(Collectors.toList());
