+++ src/main/java/wooteco/subway/dto/converter/PathConverter.java
@@ -1,4 +1,4 @@
-package wooteco.subway.controller;
+package wooteco.subway.dto.converter;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -8,11 +8,11 @@ import wooteco.subway.dto.service.PathServiceRequest;
 import wooteco.subway.dto.service.PathServiceResponse;
 
 public class PathConverter {
-    static PathServiceRequest toServiceRequest(Long source, Long target, int age) {
+    public static PathServiceRequest toServiceRequest(Long source, Long target, int age) {
         return new PathServiceRequest(source, target, age);
     }
 
-    static PathResponse toResponse(PathServiceResponse pathServiceResponse) {
+    public static PathResponse toResponse(PathServiceResponse pathServiceResponse) {
         List<StationResponse> stationResponses = pathServiceResponse.getStations()
             .stream()
             .map(stationDto -> new StationResponse(stationDto.getId(), stationDto.getName()))
