+++ src/main/java/wooteco/subway/dto/converter/SectionConverter.java
@@ -1,4 +1,4 @@
-package wooteco.subway.controller;
+package wooteco.subway.dto.converter;
 
 import wooteco.subway.dto.controller.request.SectionRequest;
 import wooteco.subway.dto.service.SectionCreateRequest;
@@ -6,12 +6,12 @@ import wooteco.subway.dto.service.SectionDeleteRequest;
 
 public class SectionConverter {
 
-    static SectionCreateRequest toServiceRequest(Long lineId, SectionRequest sectionRequest) {
+    public static SectionCreateRequest toServiceRequest(Long lineId, SectionRequest sectionRequest) {
         return new SectionCreateRequest(lineId, sectionRequest.getUpStationId(), sectionRequest.getDownStationId(),
             sectionRequest.getDistance());
     }
 
-    static SectionDeleteRequest toServiceRequest(Long lineId, Long stationId) {
+    public static SectionDeleteRequest toServiceRequest(Long lineId, Long stationId) {
         return new SectionDeleteRequest(lineId, stationId);
     }
 }
