+++ src/main/java/wooteco/subway/dto/converter/StationConverter.java
@@ -1,15 +1,15 @@
-package wooteco.subway.controller;
+package wooteco.subway.dto.converter;
 
 import wooteco.subway.dto.controller.request.StationRequest;
 import wooteco.subway.dto.controller.response.StationResponse;
 import wooteco.subway.dto.service.StationDto;
 
 public class StationConverter {
-    static StationDto toServiceRequest(StationRequest stationRequest) {
+    public static StationDto toServiceRequest(StationRequest stationRequest) {
         return new StationDto(stationRequest.getName());
     }
 
-    static StationResponse toResponse(StationDto stationDto) {
+    public static StationResponse toResponse(StationDto stationDto) {
         return new StationResponse(stationDto.getId(), stationDto.getName());
     }
 }
